protect from global scope
wrapping code in a function like this (an IIFE: "immediately-invoked
function expression") protects variable and function names from being in
the global scope. prevents potential issues with other scripts conflicting
(using the same variables and overwriting your data/functions).

it's not a high risk, but it's tricky to debug if it happens so it's a
practice to reduce that risk.